refactor: custom hook으로 분리
